requirements.txt
django
django-allauth
psycopg2
django-cors-headers
djangorestframework
django-js-asset
django-ckeditor
Pillow
django-bootstrap_v5
coverage
